Move some namespaces on Symfony event listener
