+++ src/Component/src/Symfony/EventListener/AddFormatListener.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\Symfony\EventListener;
+namespace Sylius\Resource\Symfony\EventListener;
 
 use Negotiation\BaseAccept;
 use Negotiation\Negotiator;
