+++ src/Component/tests/spec/Symfony/EventListener/AddFormatListenerSpec.php
@@ -11,13 +11,13 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Component\Resource\Symfony\EventListener;
+namespace spec\Sylius\Resource\Symfony\EventListener;
 
 use Negotiation\Negotiator;
 use PhpSpec\ObjectBehavior;
-use Sylius\Component\Resource\Symfony\EventListener\AddFormatListener;
 use Sylius\Resource\Metadata\HttpOperation;
 use Sylius\Resource\Metadata\Operation\HttpOperationInitiatorInterface;
+use Sylius\Resource\Symfony\EventListener\AddFormatListener;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
