+++ src/Component/tests/spec/Symfony/EventListener/FactoryListenerSpec.php
@@ -11,16 +11,16 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Component\Resource\Symfony\EventListener;
+namespace spec\Sylius\Resource\Symfony\EventListener;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Component\Resource\Symfony\EventListener\FactoryListener;
 use Sylius\Resource\Context\Context;
 use Sylius\Resource\Context\Initiator\RequestContextInitiatorInterface;
 use Sylius\Resource\Metadata\Create;
 use Sylius\Resource\Metadata\Operation\HttpOperationInitiatorInterface;
 use Sylius\Resource\Metadata\Update;
 use Sylius\Resource\State\FactoryInterface;
+use Sylius\Resource\Symfony\EventListener\FactoryListener;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
