+++ src/Component/tests/spec/Symfony/EventListener/FormListenerSpec.php
@@ -11,12 +11,11 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Component\Resource\Symfony\EventListener;
+namespace spec\Sylius\Resource\Symfony\EventListener;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Resource\Metadata\MetadataInterface;
 use Sylius\Component\Resource\Metadata\RegistryInterface;
-use Sylius\Component\Resource\Symfony\EventListener\FormListener;
 use Sylius\Component\Resource\Symfony\Form\Factory\FormFactoryInterface;
 use Sylius\Resource\Context\Context;
 use Sylius\Resource\Context\Initiator\RequestContextInitiatorInterface;
@@ -25,6 +24,7 @@ use Sylius\Resource\Metadata\Create;
 use Sylius\Resource\Metadata\Operation\HttpOperationInitiatorInterface;
 use Sylius\Resource\Metadata\Operations;
 use Sylius\Resource\Metadata\Show;
+use Sylius\Resource\Symfony\EventListener\FormListener;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
